requirements.txt
asgiref==3.5.0
backoff==2.1.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
dj-database-url==2.1.0
Django==4.0.3
django-cors-headers==3.11.0
django-environ==0.11.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
idna==3.4
opencage==2.0.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pytz==2021.3
requests==2.28.1
sqlparse==0.4.2
typing_extensions==4.8.0
tzdata==2021.5
urllib3==1.26.12
whitenoise==6.0.0
